backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.1.4
transformers==4.36.0
textblob==0.17.1
pytz==2023.3